added sima related info to request and changed number types in loan request

+++ GniApi/Dtos/RequestDto/LoanRequest.cs
@@ -2,13 +2,13 @@
 {
     public class Loanrequest
     {
-        public int amount { get; set; }
-        public int amountAfterCommission { get; set; }
+        public double amount { get; set; }
+        public double amountAfterCommission { get; set; }
         public int productCode { get; set; }
         public bool limitExceeded { get; set; }
-        public int percentageRate { get; set; }
-        public int comissionPercent { get; set; }
-        public int repaymentAmount { get; set; }
+        public double percentageRate { get; set; }
+        public double comissionPercent { get; set; }
+        public double repaymentAmount { get; set; }
         public int repaymentPeriod { get; set; }
     }
 }
